server_sample/requirements.txt
curlify
requests
Pillow
prettyprinter
python-dotenv